+++ CHANGELOG.md
@@ -35,6 +35,8 @@ Developers may need to change their code
 
 ### Corrected Bugs
 
+* Fixed: [leafref in new type no work in union type](https://github.com/clicon/clixon/issues/388)
+* Fixed: [Validation of YANG leafref within union does not work](https://github.com/clicon/clixon/issues/498)
 * Fixed: [leafref instance-required with state](https://github.com/clicon/clixon/issues/632
 
 ## 7.7.0